Winbond developer pre-release #4
git-svn-id: https://subversion.assembla.com/svn/fakesmc@270 958d8502-8caa-4642-8a4d-47b05a98fa7a

+++ Plug-ins/FakeSMCSuperIO/SuperIO.h
@@ -145,7 +145,7 @@ public:
 	UInt16			GetAddress() { return m_Address; };
 	Binding*		GetBindings() { return m_Binding; };
 	Controller*		GetControllers() { return m_Controller; };
-	UInt8			GetRawVCore() { return m_RawVCore; };
+	UInt16			GetRawVCore() { return m_RawVCore; };
 	bool			FanControlEnabled() { return m_FanControl; };
 	bool			FanVoltageControlled() { return m_FanVoltageControlled; };
